PCBC-54 Fix converting numbers in response
We cannot just set null-byte in payload, because the requests are in
continuous chunk of the memory and it is illegal to write after payload_len
bytes. But from another side, there no 'safe' function to convert strings to
numbers, therefore copying the payload will be the easiest way here.

Change-Id: Ifbf6e31758b73c03f541a103dacd4862f18d7806
Reviewed-on: http://review.couchbase.org/14741
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Jan Lehnardt <jan@couchbase.com>