PCBC-91: Make view support work for non _all_docs design docs.
The original implementation did only work for _all_docs and not for
other design/view combination. The problem existed, because the URI
for _all_docs is different than for all custom ones. This fix takes
those into account accordingly.

Also adapting a test case where the additional arguments for the
_all_docs design doc are treated differently as all the others.

Also rename test files so that the test ids are incremented properly
(032 was missing).

Change-Id: Ic62cdb8a2a31cdfc127dcf456bb6af76668fb543
Reviewed-on: http://review.couchbase.org/19250
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>