Remove unnecessary warning on increment/decrement.
When trying a increment/decrement with create = 0 (which is the default),
not only false is returned, but also a warning is shown. This is not
needed since false is enough information that the key doesn't exist. Also,
the warning is hard to catch when working from a user-level.

This change makes sure that only this specific information is hidden and
all other possible errors are still shown to the user.

Forward Port from 1.0.x
- needed to change LIBCOUCHBASE_* to LCB_*

Change-Id: I1cc5fa925b84a642b5126e4b871a8661e2765386
Reviewed-on: http://review.couchbase.org/20747
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Mordechai Nunberg <mnunberg@haskalah.org>