PCBC-117 Checking that expiry times are non-negative
This adds checks to the following operations to ensure that the expiry
times specified by the user are non-negative:
* add
* set
* set-multi
* replace
* append
* prepend
* cas
* increment
* decrement
* get
* get-multi
* get-delayed
* touch
* touch-multi

If the check is not successful (i.e. if the user has supplied a
negative value), a fatal error is thrown by the extension.

This is the master-branch version of an equivalent bugfix commit on
1.0.x.  During development of this patch fix, two small bugs (arginfo
handling of flags param for multi-get, initial return_value set within
touch commands) were encountered and fixed as well.

Change-Id: I08eab920c5ca51901b4973ce6669b07b7e3acee7
Reviewed-on: http://review.couchbase.org/21090
Reviewed-by: Mordechai Nunberg <mnunberg@haskalah.org>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Jackson <mjackson@omniti.com>