Fix compression test
There was multiple errors here. First of all the code didn't work
because it didn't specify the size of the compression buffer, but
the test would never work given that we always inflate object
if we've got support for the compression algorithm even if it is
disabled for the connection.

Change-Id: I3b4dfd67c1d7f4c24ad6ac2d61288363f2d3b35a
Reviewed-on: http://review.couchbase.org/21920
Reviewed-by: Mordechai Nunberg <mnunberg@haskalah.org>
Reviewed-by: Michael Jackson <mjackson@omniti.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>